+++ EigenharpMapper/Source/UI/Utility.cpp
@@ -1,9 +1,9 @@
 #include "Utility.h"
 
 juce::Colour Utility::zoneEnumToColour(Zone zone) {
-    return zoneColours[zone];
+    return zoneColours[(int)zone];
 }
 
 juce::Colour Utility::keyColourEnumToColour(KeyColour colour) {
-    return keyColours[colour];
+    return keyColours[(int)colour];
 }
